+++ packages/data/src/api/ownedTokens/ownedTokens.ts
@@ -1,6 +1,7 @@
-import { Pagination, Token } from '../../types';
+import { Pagination, ParsedDataReturn, Token } from '../../types';
 import { fetchGraphQl } from '../../graphql/fetch';
 import { ownedTokensQuery } from './ownedTokens.query';
+import { parseData } from '../../utils';
 
 export type OwnedTokensQueryResult = {
   tokens: Token[];
@@ -9,7 +10,7 @@ export type OwnedTokensQueryResult = {
 export const ownedTokens = async (
   ownerId: string,
   { limit, offset = 0 }: Pagination,
-): Promise<Token[]> => {
+): Promise<ParsedDataReturn<Token[]>> => {
   const { data, error } = await fetchGraphQl<OwnedTokensQueryResult>({
     query: ownedTokensQuery,
     variables: {
@@ -19,10 +20,9 @@ export const ownedTokens = async (
     },
   });
 
-  if (error) {
-    console.error('Error fetching token listing counts', error.message);
-    throw error;
-  }
 
-  return data.tokens;
+  const errorMsg = `Error fetching token listing counts, ${error.message}`;
+
+  return parseData<Token[]>(data?.tokens,error,errorMsg);
+
 };
