+++ packages/data/src/types.ts
@@ -1,3 +1,5 @@
+import { GraphqlFetchingError } from './graphql/fetch';
+
 // shared by data libs and not in graphql.
 export type Pagination = {
   limit: number;
@@ -15,4 +17,20 @@ export type Token = {
   media: string;
   document: string;
   animationUrl: string;
+}
+
+export interface ParsedDataReturn<T> {
+  error: null | GraphqlFetchingError;
+  data: T | null;
+}
+
+
+export type NftAttributesQueryResult = {
+  nft_attributes: Attribute[];
+}
+
+export type Attribute = {
+  attribute_display_type?: string | null;
+  attribute_value?: string | null;
+  attribute_type?: string | null;
 }
